+++ src/pages/add/add.tsx
@@ -1,15 +1,17 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 
-import { Button, Input, Title, Wrapper } from "~/pages/add-search/add-search.styles";
+import { Button, Input, Title, Wrapper } from "~/pages/add/add.styles";
 import { useInspection } from "~/providers/inspection/inspection";
 import { InspectionActionTypes } from "~/reducers/inspector/types";
 import { isEmpty } from "~/utils";
 
-export const AddSearch: React.FC = () => {
+export const Add: React.FC = () => {
   const [text, setText] = useState<string>("");
   const { inspectionService, inspectionDispatch } = useInspection();
   const navigate = useNavigate();
+
+  useEffect(() => {}, []);
 
   const handleClick = async (): Promise<void> => {
     try {
@@ -21,13 +23,13 @@ export const AddSearch: React.FC = () => {
       });
       const { data } = await inspectionService.addInspection({ keyword: text });
 
-      navigate("/searches", {
+      setText("");
+      navigate("/search", {
         state: {
           key: text,
           id: data.id,
         },
       });
-      setText("");
     } catch (error) {
       inspectionDispatch({
         type: InspectionActionTypes.error,
@@ -37,7 +39,7 @@ export const AddSearch: React.FC = () => {
 
   return (
     <div>
-      <Title>Add Search</Title>
+      <Title>Add</Title>
       <Wrapper>
         <Input
           type="text"