+++ src/reducers/inspector/types.ts
@@ -2,7 +2,6 @@ import { Inspection } from "~/entities/inspection";
 
 export enum InspectionActionTypes {
   add = "ADD_INSPECTION",
-  remove = "REMOVE_INSPECTION",
   loading = "LOADING_INSPECTION",
   error = "ERROR_INSPECTION",
 }
@@ -11,13 +10,6 @@ export interface AddInspectionAction {
   type: InspectionActionTypes.add;
   payload: {
     inspection: Inspection;
-  };
-}
-
-export interface RemoveInspectionAction {
-  type: InspectionActionTypes.remove;
-  payload: {
-    id: string;
   };
 }
 
@@ -34,6 +26,5 @@ export interface ErrorInspectionAction {
 
 export type InspectionAction =
   | AddInspectionAction
-  | RemoveInspectionAction
   | LoadingInspectionAction
   | ErrorInspectionAction;