+++ tests/pages/add/add.test.tsx
@@ -1,14 +1,14 @@
 import { render } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
 
-import { AddSearch } from "~/pages/add-search/add-search";
+import { Add } from "~/pages/add/add";
 
 describe("<AddSearch />", () => {
   it("should be able to mount <AddSearch /> component", async () => {
     // given
     const { asFragment } = render(
       <BrowserRouter>
-        <AddSearch />
+        <Add />
       </BrowserRouter>
     );